feat(*): Move most of the functions in utils.rs
